+++ frontend/src/hooks/admin/index.ts
@@ -2,4 +2,5 @@ export * from "./useAdminActivity";
 export * from "./useAdminLogs";
 export * from "./useAdminWebSocket";
 export * from "./useAdminWsOps";
+export * from "./useNavigationGuard";
 export * from "./useWsQuery";